reorganize functions of get data dir from `moseq_infer` into `moseq_train`
